requirements.txt
Flask==2.3.3
requests==2.31.0
debugpy # Required for debugging.
google-cloud-storage==2.16.0
google-cloud-secret-manager==2.19.0
python-magic==0.4.27
